fix: map lobby fields for UI
