+++ src/features/match/model.js
@@ -1,0 +1,7 @@
+export function createMatchModel() {
+  return {
+    // Match domain logic will be implemented here
+  };
+}
+
+export default createMatchModel;
