+++ src/features/match/ui.js
@@ -1,0 +1,5 @@
+export function initMatchUI() {
+  // Match UI setup placeholder
+}
+
+export default { initMatchUI };
